fixes LL-1354
