+++ package.json
@@ -46,7 +46,7 @@
     "@ledgerhq/hw-transport": "^4.55.0",
     "@ledgerhq/hw-transport-node-hid": "^4.55.0",
     "@ledgerhq/ledger-core": "^2.2.1-rc",
-    "@ledgerhq/live-common": "^5.3.0",
+    "@ledgerhq/live-common": "^5.3.1",
     "animated": "^0.2.2",
     "async": "^2.6.2",
     "axios": "^0.18.0",
